+++ lib/startback/event/ext/context.rb
@@ -1,0 +1,5 @@
+module Startback
+  class Context
+    attr_accessor :engine
+  end # class Context
+end # module Startback